requirements.txt
pandas
streamlit
plotly
numpy
altair
duckdb
dotenv
psutil
eecloud-1.5.2078.303-py3-none-any.whl